misc renamings and import sorting
