+++ src/components/Customer/MtoShipments/HHGDetailsForm.test.jsx
@@ -4,9 +4,9 @@ import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'connected-react-router';
 
-import { history, store } from '../../../shared/store';
-
 import HHGDetailsForm, { HHGDetailsFormComponent } from './HHGDetailsForm';
+
+import { history, store } from 'shared/store';
 
 const defaultProps = {
   pageList: ['page1', 'anotherPage/:foo/:bar'],
