+++ src/components/Customer/MtoShipments/NTSDetailsForm.jsx
@@ -6,15 +6,14 @@ import { Formik } from 'formik';
 import * as Yup from 'yup';
 import { Fieldset } from '@trussworks/react-uswds';
 
-import { TextInput } from '../../form/fields';
-import { Form } from '../../form/Form';
-
 import styles from './HHGDetailsForm.module.scss';
-import { RequiredPlaceSchema } from './formTypes';
+import { RequiredPlaceSchema } from './validationSchemas';
 import { fullAddressShape, agentShape } from './propShapes';
 import { formatMtoShipment } from './utils';
-import { PickupDetails } from './PickupDetails';
+import { PickupFields } from './PickupFields';
 
+import { TextInput } from 'components/form/fields';
+import { Form } from 'components/form/Form';
 import {
   selectMTOShipmentForMTO,
   createMTOShipment as createMTOShipmentAction,
@@ -147,7 +146,7 @@ class NTSDetailsForm extends Component {
           >
             <h1>Now lets arrange details for the professional movers</h1>
             <Form className={styles.NTSDetailsForm}>
-              <PickupDetails
+              <PickupFields
                 fieldsetClasses={fieldsetClasses}
                 useCurrentResidence={useCurrentResidence}
                 onCurrentResidenceChange={this.handleUseCurrentResidenceChange}
