+++ src/components/Customer/MtoShipments/NTSrDetailsForm.jsx
@@ -6,15 +6,14 @@ import { Formik } from 'formik';
 import * as Yup from 'yup';
 import { Fieldset } from '@trussworks/react-uswds';
 
-import { TextInput } from '../../form/fields';
-import { Form } from '../../form/Form';
-
 import styles from './HHGDetailsForm.module.scss';
-import { RequiredPlaceSchema } from './formTypes';
+import { RequiredPlaceSchema } from './validationSchemas';
 import { simpleAddressShape, fullAddressShape, agentShape } from './propShapes';
 import { formatMtoShipment } from './utils';
-import { DeliveryDetails } from './DeliveryDetails';
+import { DeliveryFields } from './DeliveryFields';
 
+import { TextInput } from 'components/form/fields';
+import { Form } from 'components/form/Form';
 import {
   selectMTOShipmentForMTO,
   createMTOShipment as createMTOShipmentAction,
@@ -89,7 +88,7 @@ class NTSrDetailsForm extends Component {
           >
             <h1>Now lets arrange details for the professional movers</h1>
             <Form className={styles.NTSrDetailsForm}>
-              <DeliveryDetails
+              <DeliveryFields
                 fieldsetClasses={fieldsetClasses}
                 newDutyStationAddress={newDutyStationAddress}
                 hasDeliveryAddress={hasDeliveryAddress}
